fix: moved ClipBoardIcon.kt to shared components and added simple documentation
